requirements.txt
yt-dlp==2023.11.1
requests==2.31.0
Pillow==10.1.0
Flask==3.0.0
python-dotenv==1.0.0
waitress==2.1.2
gunicorn==21.2.0
